Add external chain configuration support, fix critical EIP-155 consensus bugs, and establish consensus-critical change protocol

+++ src/main/scala/com/chipprbots/ethereum/jsonrpc/EthInfoService.scala
@@ -29,7 +29,7 @@ import com.chipprbots.ethereum.utils.BlockchainConfig
 
 object EthInfoService {
   case class ChainIdRequest()
-  case class ChainIdResponse(value: Byte)
+  case class ChainIdResponse(value: BigInt)
 
   case class ProtocolVersionRequest()
   case class ProtocolVersionResponse(value: String)
